Raise the maximum number of files MAXFILES to 9.
Note that PlotXY considers internally each function plot a file, but it
does not check that the maximum number of function plots + the maximum
number of loaded files does not overcome MAXFILES.
This can cause segfault when the number of function plots (of a given
plot) + loaded files overcomes MAXFILES

This segfault is rare. The issue should be easy to solve since the bulk
matrix with fuction plot data is dynamically allocated. Only minor GUI
data sets, will probably need to be updated to allow up to 18 function
plots irrespective of the number of loaded files.