Add keyboard shortcuts to DataSel window
Note that PlotButton is a QToolButton and therefore cannot be
autofocused. If it were a QPushButton it could be autofocused, receive
at startup setfocus() and therefore be activated just clicking on enter
on the keyboard. An example of this behaviour in TestClineChart.
Probably, in the near future the plot button (and possibly other
buttons) will be converted into a QPushButton.